+++ gonggibap/src/app/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
 import localFont from "next/font/local";
-import { Provider } from "@/app/_components/Provider";
+import { RootProvider } from "@/providers";
 import "./globals.css";
 
 const geistSans = localFont({
@@ -49,7 +49,7 @@ export default function RootLayout({
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
-        <Provider>{children}</Provider>
+        <RootProvider>{children}</RootProvider>
       </body>
     </html>
   );
